pyproject.toml
[project]
name = "diffswarm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Justin Chapman", email = "commonmodestudio@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]==0.116",
    "pydantic-settings>=2.10.1",
    "python-dateutil>=2.9.0.post0",
    "python-ulid[pydantic]==3.0",
    "sqlalchemy==2.0.38",
]

[project.scripts]
invoke = "diffswarm.tools.invoke:invoke"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py313"

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101"]

[tool.ruff.lint.pycodestyle]
max-line-length = 88

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    # may cause issues with formatter
    "COM812",
    # incompatible with D213
    "D212",
    # incompatible with D211
    "D203",
    # missing docstring
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
]
fixable = ["ALL"]

[tool.pyright]
extraPaths = ["src"]
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"

[dependency-groups]
dev = [
    "basedpyright>=1.31.0",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "ruff>=0.12.5",
]
